… caffe2/operators/instance_norm_op.cu